// src/types/auto.ts
export interface Auto {
  id?: string;
  address?: string;
  autoBranch?: string;
  autoBranchName?: string;
  autoSubBranchName?: string;
  autoBranchDoc?: string;
  autoCondition?: string;
  autoSubBranch?: string;
  autoSubBranchDoc?: string;
  autoType?: string;
  bodyType?: string;
  branchId?: string;
  code?: string;
  color?: string;
  compensationId?: string;
  countryCode?: string;
  currencyCode?: string;
  currencySymbol?: string;
  desc?: string;
  districtId?: any;
  fuelType?: string;
  geoLocation?: string;
  groupType?: string;
  locationId?: any;
  numFav?: number;
  numRenew?: number;
  phoneNumberPrefix?: any;
  position?: { lat: number; lng: number };
  price?: number;
  shareUrl?: string;
  status?: string;
  taxType?: string;
  title?: string;
  transmissionType?: string;
  urlList?: string[];
  imageFiles: File[];
  view?: string;
  year?: number;
  nearBy?: any[];
}

export interface AutoIndex {
  id?: string;
  address?: string;
  autoBranch?: string;
  autoBranchDoc?: string;
  autoCondition?: string;
  autoSubBranch?: string;
  autoSubBranchDoc?: string;
  url?: string;
  autoType?: string;
  bodyType?: string;
  branchId?: string;
  code?: string;
  color?: string;
  compensationId?: string;
  countryCode?: string;
  currencyCode?: string;
  currencySymbol?: string;
  desc?: string;
  distictId?: string;
  fuelType?: string;
  geoLocation?: string;
  groupType?: string;
  locationId?: string;
}